+++ activity/brightpay/react/changelog.md
@@ -2189,3 +2189,14 @@
 
 ---
 
+# Brightpay Development - 2025-02-21
+
+**Commit #100**
+
+- Repository: brightpay/brightpay
+- Type: web-application
+- Language: react
+- Month: 2/2025
+
+---
+